Refactoring SearchHotels, put the recessive step into a supportJS file
